+++ lib/Braintree/CreditCardGateway.php
@@ -32,6 +32,7 @@ class CreditCardGateway
     public function create($attribs)
     {
         Util::verifyKeys(self::createSignature(), $attribs);
+        $this->_checkForDeprecatedAttributes($attribs);
         return $this->_doCreate('/payment_methods', ['credit_card' => $attribs]);
     }
 
@@ -229,7 +230,7 @@ class CreditCardGateway
      * is the 2nd attribute. $token is not sent in object context.
      *
      * @access public
-     * @param array $attributes
+     * @param array $attributes (Note: $deviceSessionId and $fraudMerchantId params are deprecated. Use $deviceData instead)
      * @param string $token (optional)
      * @return Result\Successful|Result\Error
      */
@@ -237,6 +238,7 @@ class CreditCardGateway
     {
         Util::verifyKeys(self::updateSignature(), $attributes);
         $this->_validateId($token);
+        $this->_checkForDeprecatedAttributes($attributes);
         return $this->_doUpdate('put', '/payment_methods/credit_card/' . $token, ['creditCard' => $attributes]);
     }
 
@@ -275,9 +277,10 @@ class CreditCardGateway
     private static function baseSignature($options)
     {
          return [
-             'billingAddressId', 'cardholderName', 'cvv', 'number', 'deviceSessionId',
+             'billingAddressId', 'cardholderName', 'cvv', 'number',
              'expirationDate', 'expirationMonth', 'expirationYear', 'token', 'venmoSdkPaymentMethodCode',
-             'deviceData', 'fraudMerchantId', 'paymentMethodNonce',
+             'deviceData', 'paymentMethodNonce',
+             'deviceSessionId', 'fraudMerchantId', // NEXT_MAJOR_VERSION remove deviceSessionId and fraudMerchantId
              ['options' => $options],
              [
                  'billingAddress' => self::billingAddressSignature()
@@ -431,6 +434,16 @@ class CreditCardGateway
             throw new Exception\Unexpected(
             "Expected address or apiErrorResponse"
             );
+        }
+    }
+
+    private function _checkForDeprecatedAttributes($attributes)
+    {
+        if (isset($attributes['deviceSessionId'])) {
+            trigger_error('$deviceSessionId is deprecated, use $deviceData instead', E_USER_DEPRECATED);
+        }
+        if (isset($attributes['fraudMerchantId'])) {
+            trigger_error('$fraudMerchantId is deprecated, use $deviceData instead', E_USER_DEPRECATED);
         }
     }
 }